requirements.txt
azure-functions
beautifulsoup4
lxml
pandas
pygsheets
requests